Upgrade bslib package to version 0.4.2, cli to version 3.5.0, isoband to 0.2.7, purrr to 1.0.0, packrat to 0.8.1, rsconnect to 0.8.28.
Add planning_year as ENV variable

Reduce the base font size

Add new data quality criteria for existing budget contributions to ensure that the contribution year is not null.

Factorize the creation of the gt table to ensure the same configuration for sector and inter-sector review.

Add a Token to Access Intersector View in the system.

Add an Inter-Sector Review Tab that consolidates all data entered for review.